Resolve pool, navi, session, and UI conflicts while preserving single-prefab factories and CSV drop data.

+++ Assets/01.Script/12.UIManager/UIManager/01.Portrat/PortraitView.cs
@@ -16,7 +16,7 @@ public class PortraitView : MonoBehaviour, IUIViewInitialize
     {
         this.attackRating.text = attackRating.ToString("0");
     }
-    public void SetGold(int gold)
+    public void SetGold(long gold)
     {
         this.gold.text = gold.ToString();  
     }
